apvsll/requirements.txt
appdirs==1.4.3
cycler==0.10.0
h5py==2.7.0
matplotlib==2.0.2
numpy==1.12.1
olefile==0.44
packaging==16.8
Pillow==4.1.1
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
readchar==0.7
scikit-learn==0.18.1
scipy==0.19.0
six==1.10.0
sklearn==0.0
virtualenv==15.1.0
